pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personalized_trip_planner"
version = "0.1.0"
description = "Personalized Trip Planner Agent"
dependencies = [
    "langgraph",
    "langchain-openai",
    "langchain-community",
    "tavily-python",
    "python-dotenv",
    "streamlit"
]

[tool.setuptools.packages.find]
include = ["app*"]
